Currently in the state of refactoring the code. I have updated structs (mesh ,vertex and constant) to hold there own header file , this is done for the purpose of dynamically storing these components so I can reference a singular COmponent.
Shader Comp is currently in the works of updating the relevant shader functions to new local class members. This wont take long.
~
The graphics component is the same, however I will use that project as reference to how I can achieve a the dynmamic classes,.

Object Comp not yet started on.