Works ,issue was every time I called the update function , it would override the previous updated world matrix in the updateconstantbuffer function.
Based upon the order of updating the world matrix for an object. It would override the previous update instance, only regonising the transformation applied to one of the world matrix.

I solved this by instead of placing the function "updateconstantbuffer " in the update , I placed it in the draw function , doing this doesnt override, just assings the relevant the constantbuffer once , no over riding issue, further , we moved the calculated transform function in the update function , so if we have a timer moving the rotation vectors. It will update.